[TASK] Add link icon to copy anchored URL to clipboard
All headings (`H1`-`H6`) that have an `id`-tag, or whose parent does have such an `id` set, will show a little link icon in front of them, when the mouse hovers over the title. Clicking on the icon will then copy the link to the clipboard.

Tested on current versions of Firefox, Chrome, and Edge on Windows.